requirements.txt
aiogram~=2.25
openai==0.28
python-dotenv==1.0.1
